Improve error logging when hitting sources-api
When the source-api-client is handling an unsuccessful API call it
doesn't return the error code but instead raises an exception.  This was
making it extremely difficult to figure out why something was failing
because all we got was "400 Bad Request" (the Exception Class) not _why_
it was a bad request (duplicate source_name).